Rename to `extend-discussion-status-filters`
